+++ async-raft/src/storage.rs
@@ -1,5 +1,7 @@
 //! The Raft storage interface and data types.
 
+use std::error::Error;
+
 use anyhow::Result;
 use async_trait::async_trait;
 use serde::{Deserialize, Serialize};
@@ -82,6 +84,14 @@ where
 {
     /// The storage engine's associated type used for exposing a snapshot for reading & writing.
     type Snapshot: AsyncRead + AsyncWrite + AsyncSeek + Send + Unpin + 'static;
+    /// The error type used to indicate to Raft that shutdown is needed when calling the
+    /// `apply_entry_to_state_machine` method.
+    ///
+    /// This error type is only considered for the `apply_entry_to_state_machine` method as it is
+    /// the only method which is allowed to return errors normally as part of application logic.
+    ///
+    /// For all other methods of this trait, returning an error will cause Raft to shutdown.
+    type ShutdownError: Error + Send + Sync + 'static;
 
     /// Get the latest membership config found in the log.
     ///
@@ -94,6 +104,8 @@ where
     /// If the system is pristine, then it should return the value of calling
     /// `MembershipConfig::new_initial(node_id)`. It is required that the storage engine persist
     /// the node's ID so that it is consistent across restarts.
+    ///
+    /// Errors returned from this method will cause Raft to go into shutdown.
     async fn get_membership_config(&self) -> Result<MembershipConfig>;
 
     /// Get Raft's state information from storage.
@@ -102,31 +114,42 @@ where
     /// fetch the last known state from stable storage. If no such entry exists due to being the
     /// first time the node has come online, then `InitialState::new_initial` should be used.
     ///
-    /// ### pro tip
-    /// The storage impl may need to look in a few different places to accurately respond to this
-    /// request: the last entry in the log for `last_log_index` & `last_log_term`; the node's hard
-    /// state record; and the index of the last log applied to the state machine.
+    /// **Pro tip:** the storage impl may need to look in a few different places to accurately
+    /// respond to this request: the last entry in the log for `last_log_index` & `last_log_term`;
+    /// the node's hard state record; and the index of the last log applied to the state machine.
+    ///
+    /// Errors returned from this method will cause Raft to go into shutdown.
     async fn get_initial_state(&self) -> Result<InitialState>;
 
     /// Save Raft's hard-state.
+    ///
+    /// Errors returned from this method will cause Raft to go into shutdown.
     async fn save_hard_state(&self, hs: &HardState) -> Result<()>;
 
     /// Get a series of log entries from storage.
     ///
     /// The start value is inclusive in the search and the stop value is non-inclusive: `[start, stop)`.
+    ///
+    /// Errors returned from this method will cause Raft to go into shutdown.
     async fn get_log_entries(&self, start: u64, stop: u64) -> Result<Vec<Entry<D>>>;
 
     /// Delete all logs starting from `start` and stopping at `stop`, else continuing to the end
     /// of the log if `stop` is `None`.
+    ///
+    /// Errors returned from this method will cause Raft to go into shutdown.
     async fn delete_logs_from(&self, start: u64, stop: Option<u64>) -> Result<()>;
 
     /// Append a new entry to the log.
+    ///
+    /// Errors returned from this method will cause Raft to go into shutdown.
     async fn append_entry_to_log(&self, entry: &Entry<D>) -> Result<()>;
 
     /// Replicate a payload of entries to the log.
     ///
     /// Though the entries will always be presented in order, each entry's index should be used to
     /// determine its location to be written in the log.
+    ///
+    /// Errors returned from this method will cause Raft to go into shutdown.
     async fn replicate_to_log(&self, entries: &[Entry<D>]) -> Result<()>;
 
     /// Apply the given log entry to the state machine.
@@ -139,18 +162,22 @@ where
     /// specific transaction is being started, or perhaps committed. This may be where a key/value
     /// is being stored. This may be where an entry is being appended to an immutable log.
     ///
-    /// The behavior here is application specific, but errors should never be returned unless the
-    /// error represents an actual failure to apply the entry. An error returned here will cause
-    /// the Raft node to shutdown in order to preserve the safety of the data and avoid corruption.
-    /// If instead some application specific error needs to be returned to the client, those
-    /// variants must be encapsulated in the type `R`, which may have application specific success
-    /// and error variants encoded in the type, perhaps using an inner `Result` type.
+    /// Error handling for this method is note worthy. If an error is returned from a call to this
+    /// method, the error will be inspected, and if the error is an instance of
+    /// `RaftStorage::ShutdownError`, then Raft will go into shutdown in order to preserve the
+    /// safety of the data and avoid corruption. Any other errors will be propagated back up to the
+    /// `Raft.client_write` call point.
+    ///
+    /// It is important to note that even in cases where an application specific error is returned,
+    /// implementations should still record that the entry has been applied to the state machine.
     async fn apply_entry_to_state_machine(&self, index: &u64, data: &D) -> Result<R>;
 
     /// Apply the given payload of entries to the state machine, as part of replication.
     ///
     /// The Raft protocol guarantees that only logs which have been _committed_, that is, logs which
     /// have been replicated to a majority of the cluster, will be applied to the state machine.
+    ///
+    /// Errors returned from this method will cause Raft to go into shutdown.
     async fn replicate_to_state_machine(&self, entries: &[(&u64, &D)]) -> Result<()>;
 
     /// Perform log compaction, returning a handle to the generated snapshot.
@@ -161,6 +188,8 @@ where
     /// storage implementation should export/checkpoint/snapshot its state machine, and then use
     /// the value of that export's last applied log as the metadata indicating the breadth of the
     /// log covered by the snapshot.
+    ///
+    /// Errors returned from this method will be logged and retried.
     async fn do_log_compaction(&self) -> Result<CurrentSnapshotData<Self::Snapshot>>;
 
     /// Create a new blank snapshot, returning a writable handle to the snapshot object along with
@@ -169,6 +198,8 @@ where
     /// ### implementation guide
     /// See the [storage chapter of the guide](https://async-raft.github.io/async-raft/storage.html)
     /// for details on log compaction / snapshotting.
+    ///
+    /// Errors returned from this method will cause Raft to go into shutdown.
     async fn create_snapshot(&self) -> Result<(String, Box<Self::Snapshot>)>;
 
     /// Finalize the installation of a snapshot which has finished streaming from the cluster leader.
@@ -187,6 +218,8 @@ where
     /// By the time ownership of the snapshot object is returned here, its
     /// `AsyncWriteExt.shutdown()` method will have been called, so no additional writes should be
     /// made to the snapshot.
+    ///
+    /// Errors returned from this method will cause Raft to go into shutdown.
     async fn finalize_snapshot_installation(
         &self, index: u64, term: u64, delete_through: Option<u64>, id: String, snapshot: Box<Self::Snapshot>,
     ) -> Result<()>;
@@ -202,5 +235,7 @@ where
     ///
     /// A proper snapshot implementation will store the term, index and membership config as part
     /// of the snapshot, which should be decoded for creating this method's response data.
+    ///
+    /// Errors returned from this method will cause Raft to go into shutdown.
     async fn get_current_snapshot(&self) -> Result<Option<CurrentSnapshotData<Self::Snapshot>>>;
 }